CS50_Python/Unit Testing/Re-requesting a Vanity Plate/plates.py
# plates.py

def main():
    plate = input("Plate: ")
    if is_valid(plate):
        print("Valid")
    else:
        print("Invalid")

def is_valid(s):
    # Check the length of the string (6 characters maximum, 2 characters minimum)
    if not (2 <= len(s) <= 6):
        return False

    # Make sure the first two characters are letters
    if not (s[0].isalpha() and s[1].isalpha()):
        return False

    # Numbers cannot be used in the middle of a plate; they must come at the end.
    # For example, AAA222 would be an acceptable vanity plate; AAA22A would not be acceptable.
    # The first number used cannot be a ‘0’
    i = 0
    while i < len(s):
        if s[i].isalpha() == False:
            if s[i] == '0':
                return False
            else:
                break
        i += 1

    for i in range(len(s)):
        if s[i].isdigit():
            if not s[i:].isdigit():
                return False

    # No periods, spaces, or punctuation marks are allowed
    for c in s:
        if c in ['.', ' ', '!', '?']:
            return False

    # If we pass all the tests, return True
    return True
